+++ R/extract.R
@@ -74,10 +74,7 @@ extract.data.frame <- function(data, col, into, regex = "([[:alnum:]]+)",
   }
 
   # Insert into existing data frame
-  out <- append_df(data, l, match(var, dplyr::tbl_vars(data)))
-  if (remove) {
-    out[[var]] <- NULL
-  }
+  out <- append_df(data, l, var, remove = remove)
 
   reconstruct_tibble(data, out, if (remove) var else chr())
 }